+++ media-data/api/current.api
@@ -1,29 +1,48 @@
 // Signature format: 4.0
 package com.google.android.horologist.media.data {
 
-  public final class CommandMapper {
-    method public com.google.android.horologist.media.model.Command map(@androidx.media3.common.Player.Command int command);
-    field public static final com.google.android.horologist.media.data.CommandMapper INSTANCE;
+  @kotlin.RequiresOptIn(message="Horologist Media is experimental. The API may be changed in the future.") @kotlin.annotation.Retention(kotlin.annotation.AnnotationRetention) public @interface ExperimentalHorologistMediaDataApi {
   }
 
-  @kotlin.RequiresOptIn(message="Horologist Media is experimental. The API may be changed in the future.") @kotlin.annotation.Retention(kotlin.annotation.AnnotationRetention) public @interface ExperimentalHorologistMediaDataApi {
+}
+
+package com.google.android.horologist.media.data.mapper {
+
+  public final class CommandMapper {
+    method public com.google.android.horologist.media.model.Command map(@androidx.media3.common.Player.Command int command);
+    field public static final com.google.android.horologist.media.data.mapper.CommandMapper INSTANCE;
   }
 
   public final class Media3MediaItemMapper {
     method public androidx.media3.common.MediaItem map(com.google.android.horologist.media.model.MediaItem mediaItem);
     method public java.util.List<androidx.media3.common.MediaItem> map(java.util.List<com.google.android.horologist.media.model.MediaItem> mediaItems);
-    field public static final com.google.android.horologist.media.data.Media3MediaItemMapper INSTANCE;
+    field public static final com.google.android.horologist.media.data.mapper.Media3MediaItemMapper INSTANCE;
   }
 
   public final class MediaItemMapper {
     method public com.google.android.horologist.media.model.MediaItem map(androidx.media3.common.MediaItem media3MediaItem, optional String defaultArtist);
-    field public static final com.google.android.horologist.media.data.MediaItemMapper INSTANCE;
+    field public static final com.google.android.horologist.media.data.mapper.MediaItemMapper INSTANCE;
   }
 
   public final class MediaItemPositionMapper {
     method public com.google.android.horologist.media.model.MediaItemPosition? map(androidx.media3.common.Player? player);
-    field public static final com.google.android.horologist.media.data.MediaItemPositionMapper INSTANCE;
+    field public static final com.google.android.horologist.media.data.mapper.MediaItemPositionMapper INSTANCE;
   }
+
+  public final class PlayerStateMapper {
+    method public boolean affectsState(androidx.media3.common.Player.Events events);
+    method public com.google.android.horologist.media.model.PlayerState map(androidx.media3.common.Player player);
+    field public static final com.google.android.horologist.media.data.mapper.PlayerStateMapper INSTANCE;
+  }
+
+  public final class SetCommandMapper {
+    method public java.util.Set<com.google.android.horologist.media.model.Command> map(androidx.media3.common.Player.Commands commands);
+    field public static final com.google.android.horologist.media.data.mapper.SetCommandMapper INSTANCE;
+  }
+
+}
+
+package com.google.android.horologist.media.data.repository {
 
   public final class PlayerRepositoryImpl implements java.io.Closeable com.google.android.horologist.media.repository.PlayerRepository {
     ctor public PlayerRepositoryImpl();
@@ -70,17 +89,6 @@ package com.google.android.horologist.media.data {
     property public final kotlinx.coroutines.flow.StateFlow<java.lang.Float> playbackSpeed;
     property public final kotlinx.coroutines.flow.StateFlow<androidx.media3.common.Player> player;
     property public kotlinx.coroutines.flow.StateFlow<java.lang.Boolean> shuffleModeEnabled;
-  }
-
-  public final class PlayerStateMapper {
-    method public boolean affectsState(androidx.media3.common.Player.Events events);
-    method public com.google.android.horologist.media.model.PlayerState map(androidx.media3.common.Player player);
-    field public static final com.google.android.horologist.media.data.PlayerStateMapper INSTANCE;
-  }
-
-  public final class SetCommandMapper {
-    method public java.util.Set<com.google.android.horologist.media.model.Command> map(androidx.media3.common.Player.Commands commands);
-    field public static final com.google.android.horologist.media.data.SetCommandMapper INSTANCE;
   }
 
 }
