+++ media-data/src/main/java/com/google/android/horologist/media/data/mapper/SetCommandMapper.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.horologist.media.data
+package com.google.android.horologist.media.data.mapper
 
 import androidx.media3.common.Player
 import com.google.android.horologist.media.model.Command
