+++ media-sample/src/androidTest/java/com/google/android/horologist/mediasample/playback/PlaybackErrorTest.kt
@@ -18,7 +18,7 @@ package com.google.android.horologist.mediasample.playback
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.LargeTest
-import com.google.android.horologist.media.data.Media3MediaItemMapper
+import com.google.android.horologist.media.data.mapper.Media3MediaItemMapper
 import com.google.android.horologist.media.model.MediaItem
 import com.google.common.truth.Truth.assertThat
 import dagger.hilt.android.testing.HiltAndroidTest
